// src/types.ts
export interface NodeEdgeIds {
    nodes: string[];
    edges: string[];
}

export class GraphNodeSet extends vis.DataSet<vis.Node> {
}

export class GraphEdgeSet extends vis.DataSet<vis.Edge> {
}

export interface NodeEdgeSet {
    nodes: GraphNodeSet;
    edges: GraphEdgeSet;
}
// ...
export interface GraphNode extends vis.Node {
}

export interface LoadGraphOption {
    dynamic: boolean;
    centerPointX?: number;
    centerPointY?: number;
    scale?: number;
}

export interface LoadGraphOptionCallback {
    width: number;
    height: number;
}

export interface NodeCommunityEntry {
    node: string;
    community: string;
}

export interface InitData {
    nodesNum: number;
    edgesNum: number;
    autoLayout: boolean;
}

export interface CommunityData {
    communities: Community[];
    nodeMap: NodeCommunityEntry[];
}

export interface GraphEdge extends vis.Edge {

}

export class GraphNetwork extends vis.Network {
}

export interface NETWORK_OPTIONS extends vis.Options {

}

export interface FRAME_OPTIONS {
    showFaces?: boolean;
    showGroups?: boolean;
    showNodes?: boolean;
    showDegrees?: boolean;
    showEdges?: boolean;
    showLabels?: boolean;
    showTitles?: boolean;
}


export interface EVENT_ARGS_FRAME {
    mainFrame: MainFrame;
    network: GraphNetwork;
    theme: Theme;
    htmlMainFrame: HTMLElement;
}
// ...
export interface EVENT_ARGS_FRAME_SHOW_INFO extends EVENT_ARGS_FRAME, NodeEdgeIds {
    htmlInfoBox: HTMLElement;
}

export interface EVENT_ARGS_FRAME_RESIZE extends EVENT_ARGS_FRAME {
    width?: number;
    height?: number;
    oldWidth?: number;
    oldHeight?: number;
}

export interface EVENT_ARGS_RELFINDER extends EVENT_ARGS_FRAME {
    maxDepth: number;
}
export interface EVENT_ARGS_RELLIST extends EVENT_ARGS_FRAME {
    path: RELATION_PATH;
    startNodes: string[];
}

export enum FrameEventName {
    FRAME_CLEAR_ALL_FLAGS = 'FRAME_CLEAR_ALL_FLAGS',
    THEME_CHANGED = 'THEME_CHANGED',
    DESTROY_CONTROL = 'DESTROY_CONTROL',
    REMOVE_CONTROL = 'REMOVE_CONTROL',
    CREATE_CONTROL = 'CREATE_CONTROL',
    ADD_CONTROL = 'ADD_CONTROL',
    GRAPH_CONNECTED = 'GRAPH_CONNECTED',
    GRAPH_LOADED = 'GRAPH_LOADED',
    FRAME_CREATED = 'FRAME_CREATED',
    FOCUS_NODES = 'FOCUS_NODES',
    INSERT_NODES = 'INSERT_NODES',
    NETWORK_SELECT_NODES = 'NETWORK_SELECT_NODES',
    NETWORK_DESELECT_NODES = 'NETWORK_DESELECT_NODES',
    NETWORK_SELECT_EDGES = 'NETWORK_SELECT_EDGES',
    NETWORK_DESELECT_EDGES = 'NETWORK_DESELECT_EDGES',
    NETWORK_BEFORE_DRAWING = 'NETWORK_BEFORE_DRAWING',
    NETWORK_AFTER_DRAWING = 'NETWORK_END_DRAWING',
    NETWORK_DBLCLICK = 'NETWORK_DBLCLICK',
    NETWORK_CLICK = 'NETWORK_CLICK',
    NETWORK_ONCONTEXT = 'NETWORK_ONCONTEXT',
    NETWORK_DRAGGING = 'NETWORK_DRAGGING',
    NETWORK_DRAGEND = 'NETWORK_DRAGEND',
    FRAME_RESIZE = 'FRAME_RESIZE',
    RELFINDER_START = 'RELFINDER_START',
    RELFINDER_STOP = 'RELFINDER_STOP',
    RELFINDER_STARTED = 'RELFINDER_STARTED',
    RELFINDER_STOPPED = 'RELFINDER_STOPPED',
    RELLIST_PUT = 'RELLIST_PUT',
    RESULTLISTPUT = 'RESULTLISTPUT',
    IMAGEBOXOPEN = 'IMAGEBOXOPEN',
    SHOW_INFO = 'SHOW_INFO',
}